# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pulse — Agentic Maturity Assessment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /engagements/{id}/survey:
    get:
      operationId: getSurvey
      tags: [survey]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Survey" }
    patch:
      operationId: updateSurvey
      tags: [survey]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSurveyInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Survey" }
  /engagements/{id}/survey/invites:
    get:
      operationId: listSurveyInvites
      tags: [survey]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SurveyInvite" }
    post:
      operationId: createSurveyInvites
      tags: [survey]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSurveyInvitesInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SurveyInvite" }
  /engagements/{id}/survey/responses:
    get:
      operationId: getSurveyAggregates
      tags: [survey]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SurveyAggregates" }
  /survey/respond/{token}:
    get:
      operationId: getSurveyByToken
      tags: [survey]
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicSurvey" }
    post:
      operationId: submitSurveyResponse
      tags: [survey]
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmitSurveyResponseInput" }
      responses:
        "204":
          description: Submitted
...
  /engagements/{id}/scoring:
    get:
      operationId: getScoring
      tags: [scoring]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Scoring" }
    post:
      operationId: computeScoring
      tags: [scoring]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Scoring" }
  /engagements/{id}/scoring/override:
    post:
      operationId: overrideDimensionScore
      tags: [scoring]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScoreOverrideInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Scoring" }
...
components:
  schemas:
...
    Dimension:
      type: string
      enum: [tooling, measurement, process, people, governance, culture]
    Confidence:
      type: string
      enum: [low, medium, high]
...
    SurveyQuestion:
      type: object
      required: [id, section, dimension, type, prompt]
      properties:
        id: { type: string }
        section: { type: string }
        dimension: { type: string, nullable: true }
        type:
          type: string
          enum: [single_select, multi_select, likert, rank, demographic]
        prompt: { type: string }
        options:
          type: array
          items: { type: string }
        likertMin: { type: integer, nullable: true }
        likertMax: { type: integer, nullable: true }
        likertMinLabel: { type: string, nullable: true }
        likertMaxLabel: { type: string, nullable: true }
        rankCount: { type: integer, nullable: true }
        moduleKey: { type: string, nullable: true }
        conditionalOn: { type: string, nullable: true }
    Survey:
      type: object
      required: [engagementId, templateVersion, modules, questions]
      properties:
        engagementId: { type: string }
        templateVersion: { type: string }
        modules:
          type: array
          items: { type: string, enum: [security, data, platform, design] }
        questions:
          type: array
          items: { $ref: "#/components/schemas/SurveyQuestion" }
        nudgeSchedule:
          type: array
          items: { type: integer }
    UpdateSurveyInput:
      type: object
      properties:
        modules:
          type: array
          items: { type: string, enum: [security, data, platform, design] }
        questions:
          type: array
          items: { $ref: "#/components/schemas/SurveyQuestion" }
        nudgeSchedule:
          type: array
          items: { type: integer }
    SurveyInvite:
      type: object
      required: [id, engagementId, team, status, createdAt]
      properties:
        id: { type: string }
        engagementId: { type: string }
        team: { type: string }
        emailHash: { type: string, nullable: true }
        token: { type: string }
        status:
          type: string
          enum: [sent, opened, started, completed]
        createdAt: { type: string }
        completedAt: { type: string, nullable: true }
    CreateSurveyInvitesInput:
      type: object
      required: [invites]
      properties:
        invites:
          type: array
          items:
            type: object
            required: [team]
            properties:
              team: { type: string }
              email: { type: string }
    PublicSurvey:
      type: object
      required: [engagementClient, questions, status]
      properties:
        engagementClient: { type: string }
        status:
          type: string
          enum: [open, completed, expired]
        questions:
          type: array
          items: { $ref: "#/components/schemas/SurveyQuestion" }
    SubmitSurveyResponseInput:
      type: object
      required: [answers, demographics]
      properties:
        answers:
          type: array
          items:
            type: object
            required: [questionId, value]
            properties:
              questionId: { type: string }
              value:
                oneOf:
                  - { type: string }
                  - { type: number }
                  - { type: array, items: { type: string } }
        demographics:
          type: object
          additionalProperties: { type: string }
    SurveyAggregates:
      type: object
      required: [responseRate, totalSent, totalCompleted, byQuestion, byTeam]
      properties:
        responseRate: { type: number }
        totalSent: { type: integer }
        totalCompleted: { type: integer }
        byQuestion:
          type: array
          items:
            type: object
            required: [questionId, distribution]
            properties:
              questionId: { type: string }
              average: { type: number, nullable: true }
              distribution:
                type: array
                items:
                  type: object
                  required: [value, count]
                  properties:
                    value: { type: string }
                    count: { type: integer }
        byTeam:
          type: array
          items:
            type: object
            required: [team, completedCount, suppressed]
            properties:
              team: { type: string }
              completedCount: { type: integer }
              suppressed: { type: boolean }
              dimensionAverages:
                type: object
                additionalProperties: { type: number }
...
    DimensionScore:
      type: object
      required: [dimension, score, stage, confidence, evidenceIds, rationale]
      properties:
        dimension: { $ref: "#/components/schemas/Dimension" }
        score: { type: number }
        stage: { type: integer, minimum: 1, maximum: 5 }
        confidence: { $ref: "#/components/schemas/Confidence" }
        evidenceIds:
          type: array
          items: { type: string }
        rationale: { type: string }
        overrideJustification: { type: string, nullable: true }
        signalsBySource:
          type: object
          properties:
            system: { type: integer }
            survey: { type: integer }
            interview: { type: integer }
            artifact: { type: integer }
    Scoring:
      type: object
      required: [engagementId, rubricVersion, byDimension, overall, computedAt]
      properties:
        engagementId: { type: string }
        rubricVersion: { type: string }
        byDimension:
          type: array
          items: { $ref: "#/components/schemas/DimensionScore" }
        overall:
          type: object
          required: [score, stage, confidence]
          properties:
            score: { type: number }
            stage: { type: integer }
            confidence: { $ref: "#/components/schemas/Confidence" }
        computedAt: { type: string }
    ScoreOverrideInput:
      type: object
      required: [dimension, stage, justification]
      properties:
        dimension: { $ref: "#/components/schemas/Dimension" }
        stage: { type: integer, minimum: 1, maximum: 5 }
        score: { type: number, nullable: true }
        justification: { type: string }
...